refactor:코드 변경

+++ src/main/java/kr/ac/kookmin/wink/planlist/todolist/dto/TodolistDTO.java
@@ -11,13 +11,13 @@ import java.time.LocalDateTime;
 @ToString
 public class TodolistDTO {
 
-    private int todoListId;       // 투두 리스트 ID
+    private Long todoListId;       // 투두 리스트 ID
     private String content;       // 투두 리스트 내용
     private LocalDateTime createdAt; // 생성일
-    private int userId;           // 사용자 ID
+    private Long userId;           // 사용자 ID
 
     // 모든 필드를 사용하는 생성자
-    public TodolistDTO(int todoListId, String content, LocalDateTime createdAt, int userId) {
+    public TodolistDTO(Long todoListId, String content, LocalDateTime createdAt, Long userId) {
         this.todoListId = todoListId;
         this.content = content;
         this.createdAt = createdAt;
