+++ src/main/java/kr/ac/kookmin/wink/planlist/todolist/repository/TodolistRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 @Repository
-public interface TodolistRepository extends JpaRepository<Todolist, Integer> {
+public interface TodolistRepository extends JpaRepository<Todolist, Long> {
     // 특정 사용자의 투두리스트를 찾기 위한 메소드
     List<Todolist> findByUserId(int userId);
 }
